[#44] refactor : router분리
